Kill any previously running copies of the process
The leds process behaves unpredictably if several are running at once.  So kill
off any previous copies before calling ws2811_init.  This was previously done
in the CGI, but it didn't run as root, so the kill never worked.  (How did we
display any lights at all...?)  Running it in the `leds` command itself ensures
that we'll have sufficient permissions to kill previous `leds` commands, because
`leds` is installed SUID.

In Ruby this would be a one-liner, natch.